replace logrus with standard library log/slog
syslog is untested (now using slog JSON, need to figure out what the
format is supposed to be for this), iqlog doesn't support structured
logging, so only the log levels are being used from log/slog.

Note that iqlog initialises slog on [Init], so any calls to slog, or the
standard log package will route through the iqlog handler.